Cast order.status to byte; use '==' for order IDs
OrderBook: cast the ternary result to byte when setting order.status to match the field type and avoid implicit int assignment. MarketMakingStrategy: replace .equals(...) comparisons with == for activeBuyOrderId/activeSellOrderId checks (including the P&L branch) to avoid potential NPE/boxing issues and align comparisons with how IDs are represented in this codebase.

+++ src/main/java/com/hft/orderbook/OrderBook.java
@@ -167,7 +167,7 @@ public class OrderBook {
             }
         }
         
-        order.status = order.getRemainingQuantity() == 0 ? 2 : 1;
+        order.status = (byte)(order.getRemainingQuantity() == 0 ? 2 : 1);
         return trades;
     }
